Update scaling
Modify number rounding so that all numbers are displayed to three significant figures, but no more than two decimal places.
Don't round numbers until just before rendering so that values rounded down to 0.0 are distinguished from values that are not available.
Modify placeholder style so that N/A and 0.00 are displayed inline with other numbers